# frontend/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "PTSite API",
        "version": "0.0.1"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api"
        }
    ],
    "security": [
        {
            "http": []
        }
    ],
    "paths": {
...
        "/v1/holiday-calendar/{year}": {
            "get": {
                "operationId": "holiday.calendar",
                "summary": "The holidays of one year, in date order, including the ones cancelled for that year and the extras",
                "tags": [
                    "Holiday"
                ],
                "parameters": [
                    {
                        "name": "year",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Array of `CalendarHolidayResource`",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/CalendarHolidayResource"
                                            }
                                        }
                                    },
                                    "required": [
                                        "data"
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    }
                }
            }
        },
...
        "/v1/seasons/current": {
            "get": {
                "operationId": "season.current",
                "summary": "The current season: the newest one that is open and not finished",
                "tags": [
                    "Season"
                ],
                "responses": {
                    "200": {
                        "description": "`SeasonResource`",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "allOf": [
                                                {
                                                    "$ref": "#/components/schemas/SeasonResource"
                                                },
                                                {
                                                    "type": "object",
                                                    "required": [
                                                        "default_place",
                                                        "percentages"
                                                    ]
                                                }
                                            ]
                                        }
                                    },
                                    "required": [
                                        "data"
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "const": "Nenhuma temporada aberta."
                                        }
                                    },
                                    "required": [
                                        "message"
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    }
                }
            }
        },
...
        "/v1/seasons/{season}/standings": {
            "get": {
                "operationId": "season.standings",
                "summary": "Standings of the season, highest total first. Players with the same total share a rank",
                "tags": [
                    "Season"
                ],
                "parameters": [
                    {
                        "name": "season",
                        "in": "path",
                        "required": true,
                        "description": "The season ID",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Array of `StandingResource`",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/StandingResource"
                                            }
                                        }
                                    },
                                    "required": [
                                        "data"
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    }
                }
            }
        },
        "/v1/seasons/{season}/night-suggestions": {
            "get": {
                "operationId": "season.nightSuggestions",
                "summary": "Suggested dates for a new night: the next three regular weekdays, at the regular time, from today.\nDates that already have a night are left out. Only suggestions; any date can be scheduled",
                "tags": [
                    "Season"
                ],
                "parameters": [
                    {
                        "name": "season",
                        "in": "path",
                        "required": true,
                        "description": "The season ID",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Array of `SuggestedNightResource`",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/SuggestedNightResource"
                                            }
                                        }
                                    },
                                    "required": [
                                        "data"
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    }
                }
            }
        },
        "/v1/seasons/{season}/simulate": {
            "post": {
                "operationId": "season.simulate",
                "summary": "\"What if\" standings for an imagined next night. Nothing is saved",
                "tags": [
                    "Season"
                ],
                "parameters": [
                    {
                        "name": "season",
                        "in": "path",
                        "required": true,
                        "description": "The season ID",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SimulateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Array of `SimulatedStandingResource`",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/SimulatedStandingResource"
                                            }
                                        }
                                    },
                                    "required": [
                                        "data"
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationException"
                    }
                }
            }
        },
        "/v1/seasons/{season}/night-plan": {
            "get": {
                "operationId": "season.nightPlan",
                "summary": "The season planner (admins): the regular nights between two dates, every N weeks, with holidays, emendas and\nthe Carnival weekend left out and the reason given. Dates that already have a night are marked taken.\nWith `count`, the plan stops at the night that completes that many (the rounds left). Nothing is saved;\nschedule the chosen dates with POST seasons/{season}/nights/batch",
                "tags": [
                    "Season"
                ],
                "parameters": [
                    {
                        "name": "season",
                        "in": "path",
                        "required": true,
                        "description": "The season ID",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "from",
                        "in": "query",
                        "required": true,
                        "description": "First day of the plan (Y-m-d).",
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "to",
                        "in": "query",
                        "required": true,
                        "description": "Last day of the plan (Y-m-d), at most about 18 months later.",
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "count",
                        "in": "query",
                        "description": "Stop after this many planned nights, even before `to`: the rounds the season still needs.",
                        "schema": {
                            "type": "integer",
                            "minimum": 1,
                            "maximum": 100
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Array of `PlannedDateResource`",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/PlannedDateResource"
                                            }
                                        }
                                    },
                                    "required": [
                                        "data"
                                    ]
                                }
                            }
                        }
                    },
                    "403": {
                        "$ref": "#/components/responses/AuthorizationException"
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationException"
                    }
                }
            }
        },
        "/v1/seasons/{season}/calendar": {
            "get": {
                "operationId": "season.calendar",
                "summary": "The season calendar: its nights (with the winner of finished ones, the ALL IN count and the user's answer)\nand the regular nights left out because of a holiday, an emenda or Carnival, in date order",
                "tags": [
                    "Season"
                ],
                "parameters": [
                    {
                        "name": "season",
                        "in": "path",
                        "required": true,
                        "description": "The season ID",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Array of `CalendarEntryResource`",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/CalendarEntryResource"
                                            }
                                        }
                                    },
                                    "required": [
                                        "data"
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "http": {
                "type": "http",
                "scheme": "bearer"
            }
        },
        "schemas": {
...
            "CalendarHolidayResource": {
                "type": "object",
                "properties": {
                    "date": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "scope": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "holiday_id": {
                        "type": [
                            "integer",
                            "null"
                        ]
                    },
                    "cancelled": {
                        "type": "boolean"
                    },
                    "exception_id": {
                        "type": [
                            "integer",
                            "null"
                        ]
                    }
                },
                "required": [
                    "date",
                    "name",
                    "scope",
                    "holiday_id",
                    "cancelled",
                    "exception_id"
                ],
                "title": "CalendarHolidayResource"
            },
...
            "SeasonResource": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "name": {
                        "type": "string"
                    },
                    "starts_on": {
                        "type": "string"
                    },
                    "description": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "buy_in": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "rounds": {
                        "type": "integer",
                        "description": "How many nights (\"rodadas\") the season has."
                    },
                    "nights_planned": {
                        "type": "integer",
                        "description": "The season's nights that are not archived, in any state; compare with rounds."
                    },
                    "is_open": {
                        "type": "boolean"
                    },
                    "is_finished": {
                        "type": "boolean"
                    },
                    "archived": {
                        "type": "boolean"
                    },
                    "schedule": {
                        "type": "object",
                        "description": "The regular night, used to suggest and plan dates. Weekday is ISO: 1 = Monday \u2026 5 = Friday \u2026 7 = Sunday.\nevery_weeks: 1 = every week, 2 = every other week, up to 4.",
                        "properties": {
                            "weekday": {
                                "type": "integer"
                            },
                            "time": {
                                "type": "string"
                            },
                            "every_weeks": {
                                "type": "integer"
                            }
                        },
                        "required": [
                            "weekday",
                            "time",
                            "every_weeks"
                        ]
                    },
                    "default_place": {
                        "anyOf": [
                            {
                                "$ref": "#/components/schemas/PlaceResource"
                            },
                            {
                                "type": "null"
                            }
                        ]
                    },
                    "percentages": {
                        "type": "array",
                        "description": "Share of the pot for each scoring position, in whole percent.",
                        "items": {
                            "type": "object",
                            "properties": {
                                "position": {
                                    "type": "integer"
                                },
                                "percent": {
                                    "type": "integer"
                                }
                            },
                            "required": [
                                "position",
                                "percent"
                            ]
                        }
                    },
                    "nights_count": {
                        "type": "integer"
                    },
                    "updated_at": {
                        "type": [
                            "string",
                            "null"
                        ]
                    }
                },
                "required": [
                    "id",
                    "name",
                    "starts_on",
                    "description",
                    "buy_in",
                    "rounds",
                    "is_open",
                    "is_finished",
                    "archived",
                    "schedule",
                    "updated_at"
                ],
                "title": "SeasonResource"
            },
            "SimulateRequest": {
                "type": "object",
                "properties": {
                    "pot": {
                        "type": "string",
                        "pattern": "^\\d{1,10}(\\.\\d{1,2})?$"
                    },
                    "positions": {
                        "type": "array",
                        "items": {
                            "type": "object",
                            "properties": {
                                "position": {
                                    "type": "integer",
                                    "minimum": 1,
                                    "maximum": 20
                                },
                                "player_id": {
                                    "type": "integer"
                                }
                            },
                            "required": [
                                "position",
                                "player_id"
                            ]
                        },
                        "minItems": 1,
                        "maxItems": 20,
                        "uniqueItems": true
                    }
                },
                "required": [
                    "pot",
                    "positions"
                ],
                "title": "SimulateRequest"
            },
            "SimulatedStandingResource": {
                "type": "object",
                "properties": {
                    "player": {
                        "$ref": "#/components/schemas/PlayerResource"
                    },
                    "current_rank": {
                        "type": [
                            "integer",
                            "null"
                        ]
                    },
                    "current_points": {
                        "type": "string"
                    },
                    "simulated_rank": {
                        "type": "integer"
                    },
                    "simulated_points": {
                        "type": "string"
                    },
                    "added_points": {
                        "type": "string"
                    },
                    "movement": {
                        "type": [
                            "integer",
                            "null"
                        ],
                        "description": "Places moved up (positive) or down (negative); null if the player was not ranked before."
                    }
                },
                "required": [
                    "player",
                    "current_rank",
                    "current_points",
                    "simulated_rank",
                    "simulated_points",
                    "added_points",
                    "movement"
                ],
                "title": "SimulatedStandingResource"
            },
...
            "StandingResource": {
                "type": "object",
                "properties": {
                    "rank": {
                        "type": "integer",
                        "description": "Players with the same total share a rank (1, 2, 2, 4)."
                    },
                    "player": {
                        "$ref": "#/components/schemas/PlayerResource"
                    },
                    "points": {
                        "type": "string"
                    },
                    "nights_scored": {
                        "type": "integer"
                    },
                    "wins": {
                        "type": "integer"
                    }
                },
                "required": [
                    "rank",
                    "player",
                    "points",
                    "nights_scored",
                    "wins"
                ],
                "title": "StandingResource"
            },
...
            "SuggestedNightResource": {
                "type": "object",
                "properties": {
                    "starts_at": {
                        "type": "string"
                    }
                },
                "required": [
                    "starts_at"
                ],
                "title": "SuggestedNightResource"
            },
...
        },
        "responses": {
...
            "AuthenticationException": {
                "description": "Unauthenticated",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "description": "Error overview."
                                }
                            },
                            "required": [
                                "message"
                            ]
                        }
                    }
                }
            },
            "ModelNotFoundException": {
                "description": "Not found",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "description": "Error overview."
                                }
                            },
                            "required": [
                                "message"
                            ]
                        }
                    }
                }
            },
            "AuthorizationException": {
                "description": "Authorization error",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "description": "Error overview."
                                }
                            },
                            "required": [
                                "message"
                            ]
                        }
                    }
                }
            }
        }
    }
}